mine/pyproject.toml
[project]
name = "gemeni20-all"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "chromadb>=1.1.0",
    "datasets==3.6.0",
    "google-genai>=1.41.0",
    "langchain>=0.3.27",
    "langchain-chroma>=0.2.6",
    "langchain-community>=0.3.30",
    "langchain-core>=0.3.76",
    "numpy>=2.3.3",
    "python-dotenv>=1.1.1",
    "SpeechRecognition",
    "websockets",
    "markdown",
    "pygments",
    "pypdf",
    "google",
    "google-generativeai",
    "torch",
    "datasets",
    "transformers[torch]",
    "huggingface_hub"
]
